Ops

Description

This module helps the calculator deal with operations in a functional manner by providing recursive datatypes and functions that represent expressions and the different ways in which operators can be applied to them. To add custom operators just add the token you want to represent it to either UnaryOp or OpType, if it is a binary operator, add its precedence in, add its evaluation function and then finish with the character associations!

Documentation

data OpType

Token representing binary operators

Constructors

Add
Token for addition

Sub
Token for subtraction

Mul
Token for multiplication

Div
Token for division

data UnaryOp

Token representing unary operators

Constructors

Neg
Token for negation

data Expr n

Datatype representing an expression, either a constant number, a | unary operator applied to an expression, or a binary operator | applied to two expressions.

Constructors

Const n
A constant number

Un UnaryOp (Expr n)
A unary operator applied to an expression.

Op OpType (Expr n) (Expr n)
A Binary operator applied to two expressions.

opTypePreced :: OpType -> Int

A function associating each binary operator to its precedence.

opTypeEval :: Floating n => OpType -> n -> n -> n

A function that applies a binary operator to two numbers.

opChar :: OpType -> Char

A function that associates each binary operator to its symbol.

assocOpType :: Char -> OpType

A function that associates a character to a binary operator.

assocUnSymb :: String -> UnaryOp

A function that associates a symbol to its unary operator.

opChars :: [Char]

The set of all chars that are associated to a binary operator.

unOpSymbs :: [String]

The set of all strings that are associated to a unary operator.

eval :: Floating n => Expr n -> n

A function that recursively evaluates an expression to the fractional it represents.