Parser

Description

This module's purpose is to parse a string into a valid expression. To acheive this, a system of interpretable tokens is introduced that allows the parse function to serially compile a token sequence into an expression.

Documentation

parse :: Read n => String -> Expr n

Parses a string into an expression. For nums and parseable values | attached to a unary operator, parsing is done simply recursively, but for | more complicated strings, the string is converted into tokens and the | tokens are compiled into the expression it represents.