StringFuncs

Description

This module deals with strings and the different ways to parse and verify them for the calculator! The most important functions done are parentheses cutting out (either simply cut out or spanned), chunking into tokens, and validating whether or not is actually represents a calculable string.

Documentation

trimOuterParens :: String -> String

Trims the outer parentheses from a string if they are present.

chunkByPredicate

Arguments

:: (a -> Bool)
Predicate to chunk by.

-> [a]
List to chunk.

-> [[a]]
Chunked list.

Chunk list by a predicate w/o any triggers.

breakUpString :: String -> [String]

Breaks up a string into number, operator, or parenthetical sections.

strIsCalcValid :: String -> Failable Bool

Determines if a string is a valid calculator string. A string is valid | if and only if its parentheses are balanced, its operators are valid | and there are no unknown characters present in it. | TODO: Make noWeridStuff more accurate by checking if any non-operator | strings are present, not just characters.

strToTokenChunks :: String -> [String]

Chunks a string into chunks that are calculator interpretable tokens.

spanWhileNotUnOp :: String -> (String, String)

Spans from a string while its accumulator is not a valid unary operator.

isNum :: String -> Bool

Determine if a string is a number

startsWithOps :: String -> Bool

Determine if a string starts with unary operators.

isOp :: String -> Bool

Determine if a string is an operator

isExpr :: String -> Bool

Determine if a string is an expression